requirements.txt
# Core document processing
pymupdf>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.2.0
docx2txt>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# Utilities and validation
pydantic>=2.5.0
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0
psutil>=5.9.0
watchdog>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Testing and validation
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
